requirements.txt
-i https://pypi.org/simple
cffi==1.17.1; platform_python_implementation != 'PyPy'
charset-normalizer==3.4.2; python_version >= '3.7'
cryptography==45.0.4; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
lxml==6.0.0; python_version >= '3.8'
pdfminer==20191125; python_version >= '3.6'
pdfminer.six==20250506; python_version >= '3.9'
pycparser==2.22; python_version >= '3.8'
pycryptodome==3.23.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
python-docx==1.2.0; python_version >= '3.9'
typing-extensions==4.14.0; python_version >= '3.9'
requests==2.32.4; python_version >= '3.9'
python-dotenv==1.1.1; python_version >- '3.9'
